refactor: HomeGalleryをデータフック・サブコンポーネントに分割
