+++ src/components/home/GalleryStates.tsx
@@ -1,0 +1,33 @@
+// Loading State
+export function LoadingState() {
+  return (
+    <div className="space-y-4">
+      <div className="aspect-[4/3] bg-gray-200 rounded-2xl animate-pulse" />
+      <div className="grid grid-cols-3 gap-1">
+        {Array.from({ length: 6 }).map((_, i) => (
+          <div key={i} className="aspect-square bg-gray-200 animate-pulse" />
+        ))}
+      </div>
+    </div>
+  );
+}
+
+// Error State
+interface ErrorStateProps {
+  message: string;
+  onRetry: () => void;
+}
+
+export function ErrorState({ message, onRetry }: ErrorStateProps) {
+  return (
+    <div className="bg-red-50 rounded-2xl p-6 text-center">
+      <p className="text-red-600 mb-4">{message}</p>
+      <button
+        onClick={onRetry}
+        className="px-4 py-2 bg-red-100 text-red-700 rounded-full text-sm font-medium hover:bg-red-200 transition-colors"
+      >
+        再試行
+      </button>
+    </div>
+  );
+}
